rename trpc helper file
