+++ src/api/routes/v2/router.ts
@@ -1,4 +1,4 @@
-import { router } from "./trpc";
+import { router } from "./trpcHelpers";
 import { authRouter } from "./routers/auth";
 
 // Main tRPC router - this is what gets exported to the types package
